Implemented main screen and salary calculator, I'm so done with this shit

+++ app/src/main/java/com/example/seamstressautomatization/data/entities/Stuff.kt
@@ -9,5 +9,5 @@ data class Stuff(
     @PrimaryKey(autoGenerate = true)
     var id: Int = 0,
     var stuff_name: String,
-    var salary: Int,
+    var salary: Float,
 )
